Read from namelist and streams in analysis funcs
This merge changes ocean.ohc_timeseries,
ocean.sst_timeseries, seaice.modelvsobs and seaice.timeseries to use
namelist and streams files to read config options, and determine the
name of mesh, input and output files to be read in.

The names of the namelist and streams files are now read from the config
file, and have therefore been added to config.analysis.

Note: in the case of ocean.ohc_timeseries, the input file, not the
mesh file is being used because we need a dycore specific fields,
refBottomDepth, that is not guaranteed to be in the mesh file.