Add startDate and endDate to StreamsFile.readpath
This merge adds arguments for a startDate and endDate when reading
a path from a StreamsFile object.  readpath now returns a list of
files matching the template, rather than the template itself (making
use of shared.io.utilities.paths).  If a startDate and/or endDate
is supplied, the list of files includes only files after the start
date and before the end date, with a buffer of one output interval
on either side to make sure no relevant data is missed.

This will allow analysis to be performed on a subset of the output
(e.g. for debugging purposes).