Add to_datetime and to_timedelta to the Date class
The to_datetime method can be used (optionally with an offset year)
to convert a Date to a datetime.  The date is clamped to a valid
range supported by numpy's datetime64[ns], used by xarray and
pandas.  This method will be useful for computing datetime bounds on
time-series data sets.

The to_timedelta method converts a Date object with
isInterval==True to a timedelta object.