Add has_stream and find_stream to streams file reader
has_stream checks if a given stream is present in a streams
file.  The ability to check for a given string is useful in
determining which stream name is appropriate for a given MPAS-O
version.

find_stream uses has_stream to find which stream in a list of
possibilities is present in the file.

Also, a few changes have been made to make the module
namelist_streams_interface.py PEP8 compliant.