Add support to mpas_xarray for a variable map
The new function map_variable(...) finds the MPAS dycore variable
in a given dataset that corresponds to a given mpas_analysis
variable name.

The new function rename_variables(...) renames all vriables in a
given dataset from MPAS dycore names to mpas_analysis names if they
are found in the variable map.

An argument for a variable_map has ben added to preprocess_mpas(...).
This map, if present, is used to fine the appropriate time
variable(s) in the dataset and to rename variables (other than the
time variable(s)) to their mpas_analysis names.