Add stream and var maps to run_analysis and to OHC
Stream and variable maps for use in the ocean analysis are now
imported as part of initializing the analysis.  The maps are
passed on to the OHC time-series analysis function.

The OHC time-series analysis has been updated to used the maps
to find the appropriate time-series stream and to map variable
names to more simplified names that will become the MPAS-Analysis
standard names.  These more simplified names are then used in
the remainder of the analysis.

Also, ohc_timeseries.py has been made PEP8 compliant.