Returns error for invalid path in streams file
Ensures that paths read in streams files are valid.
An error is reported if the path in the streams file
does not exist.

This is important for analysis of ACME files that have
been moved from their original locations because
the path in the streams files may no longer be relevant.