Move parts of ocean modelvsobs to shared climatology
Eventually, code for creating a climatology should be made general
enough to be useful across components.  This reorganization is
also intended to make it easier to change the approach used for
performing horizontal interpolation, which is not isolated to
the shared climatology module.

The resolution of the comparison grid for ocean observations
has been changed to 0.5 deg from 1 deg to be consistent with
sea ice climatologies and prior agreement on what the comparison
grid should be.

constants module has been made PEP8 compliant.

Additional sequences of months needed by the sea-ice climatology
calculations have been added to the monthsDictionary in constants.