Include config.default in all runs
Change all other config files to include only options that are
different from the default (or which a user would likely change).

This greatly reduces the changes required when the default config
file is modified in the future in ways that do not impact machine-
and run-specific config files.

Modify the README base on new config.default

Add warning when config.default not found