Add support for running tasks in parallel
This is controlled with the 'parallelTaskCount' option in the
'execute' section, which defaults to 1 (serial) and the
'commandPrefix' option in the same section, which can be used to
add a prefix such as 'srun -n 1' or a path to python if needed.

If there is more than one task to run and parallelTaskCount > 1,
several tasks run in parallel, and their output is stored in log
files instead of being output to stdout/stderr.