Add autoclose to xarray.open_mfdataset call
We set autoclose to true if the number of files is too close to
the maximum number of files that can be open at a given time.

A new option, autocloseFileLimitFraction, has been added that
determines what fraction of the system limit of open files are
allowed.  The default is 50%.

Also, a bit of PEP8 clean up.