Update climatologies to use start and end times
There is no longer an advantage to computing monthly climatologies
first, then seasonal climatologies (as was done in several analysis
tasks), so seasonal climatologies are now computed directly.

The MOC time average is now computed using compute_climatology
(rather than just .mean('Time')) so that the number of days in a
month is properly accounted for.