Update Nino34, climatology maps w/ climo caching
Yearly or multi-year climatologies are first computed and cached.
Then, a climatology over the full requested range is computed and
cached.

This appears to make memory usage quite manageable and performance
reasonable, in particular for the MOC.. It should also
substantially derease execution time when analysis is performed
multiple times on a simulation in progress.

Adds maskVaries option for computing climatologies.  Most MPAS data
should have maskVaries=False, since the mask cannot change in time.
It is not clear whether sea ice thickness and concentration might
best have a mask (which varies in time) but previously sea ice
was not masked and this is left unchanged by this merge.