Fix errors with caching climatologies
In cases where the single cached climatology file has the same
name as the aggregate climatology file, we don't need to aggregate.
We also want to make sure to close open data sets to prevent
conflicts when writing to files of the same name (though this
situation hopefully is avoided in any case).

Add a helper function for creating a correct file suffix and range
of years depending on whether 1 year or several.