Addresses fixes for MOC for 18to6
Use of *.compute() was dangerous because it required
xarray / dask to store arrays in memory, which is problematic
for large problem sizes like the 18to6.  For the MOC
calculation for the 18to6, use of maxChunkSize is required to
ensure that the large arrays fit into memory for the climatology
calculation.  However, for smaller problem sizes (60to30)
use of maxChunkSize will result in a performance penalty relative
to this value being None because computations will be faster if
arrays fit into memory.