Remove remapper from get_mpas_climatology_file_names
There are cases where climatology file names are needed but
remapping will not be performed.  In such cases, only the MPAS mesh
name needs to be provided, not the comparison grid name.  This
merge makes the comaprison grid name an optional argument and
removes the requirement that a remapper be provided to
get_mpas_climatology_file_names.