Change climatology to classes
Change functionality of the climatology module to be handled by
3 new classes:
Climatology - the base class
MpasClimatology - for computing, caching and/or remapping MPAS climatologies
ObservationClimatology - for computing and/or remapping observational
                         climatologies

These classes know more about tasks than the previous standalone functions,
and can therefore automatically handle much more of the process of setting
up climatologies.  This makes the tasks themselves shorter and cleaner, and
hopefully easier to understand.

Functionality has been added to AnalysisTask to cache the times only from
a given stream.  This is useful for later determining which files need
to be opened when computing a given climatology or time series and which
have already been cached.  Climatology caching in the MpasClimatology
class has been updated to use the time cache.  Note: the time cache is a
dictionary, and thus has been stored in a python ".pickle" file, since
NetCDF doesn't support python dictionaries.

Since many tasks depend on the results from the cached times, a task has
been added to cache these times before calling other tasks.  Each analysis
task now has a list of prerequisite tasks that must run before it can be
run.  In addition to supporting caching times, in the future this functionality
could be used to build tasks out of multiple subtasks (e.g. one to compute
the climatology and one to do the plotting).  This would be particularly
useful if multiple tasks depend on the same climatology (which is not currently
the case).

All analysis tasks that use the climatology module have been updated to
use the new classes.

All tests related to climatologies and the AnalysisTask base class have been
updated.

The functions "read" and "create" in MeshDescriptors have been made static
functions that return an instance of the class, saving the need to call the
empty constructor before calling these methods.

mpas_xarray has been updated so that a coordinate can be passed to the variable
list.  This is useful if the only information needed from a data set is the
Time coordinate (as is the case for time caching).

Cleanup in config.default:
* rename regridded to remapped
* remove overwrite flags (we assume they are false and the better
  way to handle this is to delete the cache files)