Remove lock around mapping files
The lock caused problems on some file systems.  Instead, to prevent
file-name conflicts between processes, mapping files created with
parallel tasks are now given unique names including the task name.
This is slightly inconvenient for copying them to a common location
for later use but avoids needs for locks or interpocess communication.

The file names are the same as before (and redundant mapping files
are created only once) if running in serial (prallelTaskCount = 1)