Fix path to namelist files
The path is now be relative to the baseDirectory, not to the
runDirectory, as specified in the comment in config.default.

This error was not detected in MPAS-Analysis tests because they
all use the same baseDirectory and runDirectory, but the same
is not true of A-Prime.  Probably a standard test that simulates
ACME short-term archiving is needed.