Remove file padding when getting file list from stream
Previously, an attempt was made to add one extra file before and
after the requested range of dates to handle cases where dates in
the files were not consistent with dates in the file name.  This
should no longer be a problem in MPAS, the buffer files are not
needed.

The extra files cause occasional confusion that would be nice to
avoid.  If a user asks for a range of data that is not available,
but one of the extra files *is* available, the result is an empty
data set rather than an error message that the no files are
available in the requested range.

Test updated accordingly