Add option ncclimoParallelMode; default is serial
This merge adds a config option ncclimoParallelMode in the execute
section.  The default is "serial", but it can also be set to "bck"
on machines with enough RAM to handle the 12 background processes
for computing monthly climatologies simultaneously.

This should be tested on different machines and for different run
sizes, and the machine-specific config files should be updated
accordingly.