Remove alpha8, beta0 runs on edison; add G-QU240
The switch to ncclimo makes it difficult to accommodate the variable
name mapping required to support older ACME versions (before beta1).
The simplest solution is to not support these older versions any longer
and to move away from mapping (which is confusing to developers).

The new G-QU240 run is small and helpful for debugging MPAS-Analysis.