Add write_netcdf function with finite _FillValue
With this merge, xarray data sets can be written out with finite
_FillValue using the shared.io.write_netcdf function.  By default,
the fill values are the defaults from the netCDF4 package.

All climatologies have been updated to use write_netcdf.

ncremap is now called with the --renormalize flag so that only
non fill values are used in the resulting output.