Add support for prerequisite tasks
Each analysis task now has a list of prerequisite tasks that must
run before it can be run.  In addition to supporting caching times,
in the future this functionality could be used to build tasks out
of multiple subtasks (e.g. one to compute the climatology and one
to do the plotting).  This would be particularly useful if
multiple tasks depend on the same climatology (which is not
currently the case).

A little bit of clean up has been done on some __init__.py files
and on shared.io.utility