Create tasks with create_tasks method
This allows a call to create_tasks for a given class to generate
not just one but multiple tasks.  This could be useful for tasks
that can be broken into multiple subtasks (each of which can be an
AnalysisTask).  It is also useful for analysis involving a number
of independent plots (e.g. multiple seasons or depths) that can be
handled by separate tasks in parallel.