Fix confusing None return from getExpression
The getExpression method from the MpasAnalysisConfigParser was
returning None when the config option was not present.  This was
confusing when the option was not intentionally left out, leading
to non-intuitive error messages.  Now, getExpression will raise
a NoOptionError (as do all get* methods) when the option is missing.

To accommodate the special needs of chunking, the chunking option
of the generalize reader has been set to None if no chunking is
specified in the config file.

CI tests have been updated accordingly.