Add antarctic comparison grid to ocean clim. maps
SST, SSS and MLD can be plotted on this grid by adding 'antarctic'
to the comparisonGrids list.

A plotting routine for data on a ploar grid has also been added,
including support for both index and norm color maps (the latter
will be used for melt rates and Antarctic temperatures at various
depths).