Add analysis tasks for plotting SOSE T and S
They can be plotted over various seasons and at various depths.

A few small additions and fixes to MpasClimatology and
ClimatologyMapOcean are required to accommodate these new tasks.
Notable is the addition of a function for post-processing the
regridded climatology (needed so the fields can be indexed at
the requested depth).