Renormalize climatologies also when using ncremap
With the change to write_netcdf, ncremap now extrapolates with
bilinear remapping but does not normalize by default.  This leads
to values closer to zero than expected in extrapolated regions.
When we do not use ncremap, we renormalize these values based on
only the weights of unmasked source grid cells.  With this merge,
we do the same when using ncremap.