Add module for generating results webpages
Classes MainPage and ComponentPage can be used to generate
webpages to display results.  The contents of the templates
subdirectory are used to construct the pages.

The pages rely on xml files associated with each image being added
to the page, and they require that images are added to the page in
the order they should appear.  The xml files do not yet exist and
the mechanism for traversing images in desired order has not yet
been fleshed out.