Update climatology start/end year from file names
Switch to updating start and end year for climatologies based on
available file names, rather than the times in the data sets.
This is needed because we need to know the names of the plots that
each task will produce without actually opening the associated
data set, and the start and end year are included in many
climatology file names.