Switch to using multiprocessing instead of subprocess
In order to support better data availability between tasks (e.g.
if tasks depend on data from the setup of other tasks), tasks are
now subclasses of multiprocess.Process.  This work is needed as
a precursor to adding support for subtasks (e.g. for producing
cliamtologies, or for generating each plot from a task).

Logging is also altered so that each task has a logger, to which
info, errors and warnings should be logged.

Because multiprocess.Process already uses a method called run(),
the run() method of AnalysisTask and its dervied classes had to
be renamed to run_task().

Job scripts have been updated to run on 1 node. The default
ncclim mode is background (12 threads)

The path to @czender's NCO installation has been removed from
the edison job script.  It is no longer needed for the latest
NCO and MPAS-Analysis, which includes a flag to ignore
@czender's path on LCF machines.

The LANL jobscript has been updated to have 1 Open MP thread
instead of 2.  This is needed (as on NERSC) to prevent NCO
problems when running with the default 2 threads. (Not sure why.)