Add --list and --purge command-line options
--list will list all available analysis tasks (useful for deciding
what to include in the --generate flag).  No config file should be
needed. The tags associated with each task are also printed.

--purge will delete the base output directory so the analysis can
be run cleanly at that location.  A config file must be supplied
with section [output] and option baseDirectory giving the location
to be deleted.  The user is told whether the location was deleted.

Updated the README and documentation to include these changes.