Add setup_only and html_only flags
These are used to skip the run (and possibly the html-generation)
phases of MPAS-Analysis.  A reason to do only the setup phase
would be to create maping files without invoking MPI (which seems
to cause problems with ESMF_RegridWeightGen).  A reason to do
HTML-generation without running would be if the analysis has
already been performed (e.g. in several steps) and the only thing
that is needed is to re-generate the HTML for all tasks that were
run.