Fix time axis in Nino observations for numpy 1.13
There is a problem with using numpy.datetime64 with the latest numpy
as an x axis in plots.  Therefore, time is now expressed as days
since 0001-01-01 (same as for model output).

A bit of layout clean-up was required to accommodate the fix.

A bit of code and docstring clean up was performed along the way.