Remove `iter` prefix when iterating through dicts
In python 2, calling just `.items()`, `.values()`, or `.keys()` returns
a copy of dictionaries list of key-value pairs, values, or keys,
respectively. This results in extra memory overhead when used to iterate
over a dictionary and instead it was common to use the `.iteritems()` methods
to generate an iterator.

In python 3, `.items()` now returns an iterator-view, and the `iteritems()`
method has been removed.

Because python 3 is the current target, using `.items()` and taking the
memory hit in python 2 is preferable over using ifs or packages like six to
provide efficient iteration across both python 2 and 3.