Add subtask for remapping observations
This merge creates a generic subtask for remapping observations to the
comparison grid that can be used by both ocean and sea-ice tasks.

Tasks for ocean and sea-ice climatology maps are broken out into
their own files, each of which defines the task and a subtask for
reading in the observational data set and defining the source
mesh/grid for the observations.

In each case, the main task now simply spawns subtasks for
remapping model results and observations and for plotting the
results and is otherwise empty.