Add ClimatologyMapAntarcticMelt task
This task plots climatologies of Antarctic melt rates against
observations from Rignot et al. (2013).

To support masking the melt rate field only where land-ice is
present and renormalizing the field to change to units of m/yr, a
capability to customize climatologies during the remapping process.

A new directory for the melt observations has been added, along
with a new section in the config file for this task.