Change obs to ref in 3 plot functions
The routines are plot_polar_comparison, plot_global_comparison
and plot_polar_projection_comparison.  This chanrge has been made
because the second panel may be used to plot a reference run
instead of observations, depending on the particular analysis
configuration.

Also, each routine plots only 1 panel instead of 3 if there is no
refArray.  This is useful for plots with no observations or
reference run to compare against.