Add ref. climatology tasks
These tasks just make sure the climatology and remapped climatology
files exist in the reference run.

The task name of MpasClimatolgyTask can optionally be set, so that
the reference run's task doesn't have the same name as the equivalent
for the main run

Similarly, the remapping subtask needs a way to set its subtask.