Add ref. run support to TimeSeriesAntarcticMelt
This also requires writing out the aggregated melt time-series data
set, so that this data set can simply be read in from the reference
run.  Because it is easy to include, this merge also loads the
time series data for the main run if the data has already been
cached.