Add pot. and neutral density to SOSE process. script
Also, switch from "z" to "depth" in SOSE output, since it is
positive up.  This will mean having to replace SOSE data on all
machines once the update is in place.

To facilitate potential density, zBot was added as a coordinate.

There were issues with velocity magnitude on the original lat/lon
grid (the 2 components aren't on the same grid) so this is only
computed on the Antarctic grid.