Make "--purge" smarter about what it deletes
Delete only known subdirectories, not just the whole base tree.
This should prevent deleting other contents that might be in the
base directory that shouldn't be purged and will also purge
"subdirectories" with absolute paths like the html.

Also, add more discussion of when to use and not use the --pruge
option.