Fix handling of masking in sea-ice plots
Where input data (both model and obs/ref) is NaN, the plots should
be masked and are now handled with a masked array as on the ocean
side.  Where either model or obs is masked, the diff is now masked.
Only after computing the diff is the other threshold applied to
model and obs/ref (e.g. zero value).  This is because we don't
want the difference to be masked based on where either model or
obs is zero, only where either is invalid (NaN).