Update example BGC config files
After the refactor, a particular variable should be excluded from
the list of BGC variables rather than trying to disable tasks
with that field as a tag (because the latter disables all BGC).

The paths to observations, mapping files and region masks also
need to be updated to the maintained locations.