Update time bounds *before* analysis tasks are created
This allows analysis tasks to use the time bounds to create
subtasks (e.g. to process parts of a time series in separate tasks)
which is not possible if time bounds are inconsistent with
available output files at init.

It also means no one task has to update these bounds, which may
avoid problems with different behavior depending on which set of
tasks is run.