Fix depricated warning about list indices vs. tuples
In the future, indices will need to be tuples to best to fix this
before it becomes a problem.

Tuple indexing that produces a single value gives that value as
a float, not a 1-element numpy array (as list indexing does), so
conversion to numpy.array is needed in several places.