Add diagnostics config section
This includes a baseDirectory and the mapping and region-masks
subdirectories, using the standard directory layout as on the
E3SM diagnostics server.

The various *Observations sections have also been modified to
reference this new standard base directory so users will need to
change very little to set up a new config file.

Many tasks and quite a bit of documentation needed to be updated
to handle these seemlingly minor changes.

Example config files will be updated in a separate commit.