Check for enough data in each anomaly task and nino34
The anomaly tasks cannot meaningfully run on a year or less of
data.  They now all check during setup to see if there is sufficient
data an fail with a warning if not.  (The reason for the failure
is not given unless the --verbose tag is used.)

Add text telling users to get more info with the --verbose tag

Don't overwrite MOC time series if the yearly and total data are
the same