Update inset with red starting and green ending dots
Add colors to tanset y axes to indicate start and end as well.

Vertical sections now provide the axis (or axes) as return values
along with the figure and the super-title.

The super-title for transects has been nudged to the left to make
room for the inset.  Titles have also been shortened.