Fix documentation
The `docs` folder was being included in the package, and to avoid
this some code has moved into the mpas_analysis package.

A table with current conda-forge status, etc. was being included in
the Quick Start that messed things up.

The setup.py now excludes the docs directory (and the version number
has been corrected)

A bit of cleanup to RTD as well.

Fix the docstring for insets.