Add customDirectory for diagnostics
This allows a user to supply their own observations, mapping files
or region masks (without putting them in a shared diagnostics
directory). This is appropriate when working with custom meshes or
testing new observations or regions that aren't (yet) standard in
MPAS-Analysis.

In the process of testing, I found bugs in timeSeriesOceanRegions that I
have fixed here.  They relate to the names of subtasks, which were not
unique when a second region group is defined.