Compute and store regional T, S for T-S diagrams
The regional T and S depends on masking from the restart file,
meaning it cannot be computed for a control run (for which a
restart file is not available during main vs. control analysis).
This means that it needs to be stored in a file to be read in
and plotted, which also should significantly improve efficiency
when re-running analysis or performing main vs. control runs.