Update docstrings to get rid of recent warnings
In a few cases, lists of items were being referenced inside of
a string literal, which seems to cause trouble since sphinx/napoleon
are also trying to turn them into a string literal.

There is no reason to generate documentation for the
`MpasAnalysisConfigParser` class, since the constructor and most
methods are inherited from `ConfigParser`.