Use cartopy coastline by default in  polar projection plots
The user can optionally switch back to the old behavior of using
the MPAS coastline.

Lat/lon contours have been added (partly to cover up a seam in the
coastline geometry).

The default color for continents is now the lighter gray used for
polar projections, and this also applies to global plots.