Fix hovmoller plot comparison
Sometimes, runs have slightly different Time values in
timeSeriesStatsMonthly (seemingly related to restarts). When
this happens, the hovmoller differences are only at times that
match exactly, leading to a diff field with the wrong size.

This merge fixes the issue by first checking to make sure the
two runs start and end at (approximately) the same time, then
using the time from the "main" run for the "reference" run, too.