Add an alternative xarray approach to computing convservation time series
This alternative approach makes sure the constents are unique (and
sorted) in the xtime variable.

To reduce redundancy between the new approach and the old ncrcat
approach (retained for comparison and in case the xarray approach
is not performant), a new helper function for determining if we
want to append to the existing datasets added.