Add task for plotting maps of wind stress curl
The wind stress curl is first computed on the native grid (at
vertices) and then is remapped to comparison grids.

The reconstruciton of the wind stress at edges is computed with
a utility funciton `vector_cell_to_edge_isotropic()`, analogous
to the subroutine with the same name in the MPAS framework.